requirements.txt
Flask
pymysql
flask-cors
python-dotenv
gunicorn
bcrypt